build: land brush_edges + 6 database + 2 devgui + aim_target_mp
POSIX build: 10 more files in. Switch mirrors all except brush_edges
(deferred — GCC -Warray-bounds on hex-rays-style indexing).

common/brush_edges.cpp: clang-only for now. Switch GCC still flags
`(*xyz)[3*k]` indexing because xyz is declared as float(*)[3] but
the actual data is a flat float buffer. Real fix wants a signature
refactor.

database/ (6):
- db_assetnames, db_auth, db_memory (these became clean once the
  d3d9.h include in r_buffers.h got guarded).
- db_stream, db_stream_load, db_stringtable_load — needed KISAKHACK
  bridges where the upstream code stores 64-bit pointers in 32-bit
  cells (uint8_t* alignment via uintptr_t, offset-to-pointer conversion,
  SL_GetString result reinterpret). All paths are asset-loader code
  that won't actually run until DB_LoadXAssets/FastFile is real; the
  hacks just keep the link going. Documented in docs/RISKS.md.

devgui/ (2): devgui_input, devgui_util compile clean.

aim_assist/ (1): aim_target_mp. One [[maybe_unused]] on a debug
DObj* and one unused-label LABEL_25 removed.

gfx_d3d/r_buffers.h: d3d9.h include now guarded by _WIN32. The 9
IDirect3D* opaque forward-decls in kisak_compat.h keep the rest of
the header parseable on POSIX/Switch.

posix/kisak_compat.h:
- OVERLAPPED type added (struct _OVERLAPPED with the Win32 nominal
  layout — async I/O state, never actually used on POSIX).
- EnterCriticalSection/LeaveCriticalSection/InitializeCriticalSection/
  DeleteCriticalSection no-op shims (added earlier this session).

posix_backbone_stubs.cpp additions (~30 entries):
- CL_ClearKeys, Key_IsDown, CG_TraceCapsule, CG_DObjGetWorldTagPos.
- FX_GetClientVisibility, BG_EvaluateTrajectory, DevGui_Toggle.
- R_AddCmdDrawText/StretchPic/StretchPicRotateXY/QuadPic,
  R_TextHeight/Width, R_AllocStatic{Vertex,Index}Buffer,
  R_FinishStatic{Vertex,Index}Buffer, R_FreeStatic{Vertex,Index}Buffer,
  R_UnlockVertexBuffer, R_UnlockIndexBuffer.
- DB_LoadXFileData, PMem_Alloc, PMem_GetOverAllocatedSize.
- SL_GetString, SL_AddUser, RadiusFromBounds (real impl).
- g_assetNames + varXAssetList globals (opaque storage).

+++ src/database/db_stream.cpp
@@ -81,7 +81,9 @@ uint8_t *__cdecl DB_GetStreamPos()
 uint8_t *__cdecl DB_AllocStreamPos(int32_t alignment)
 {
     iassert(g_streamPos);
-    g_streamPos = (uint8_t *)(~alignment & (uint32_t)&g_streamPos[alignment]);
+    // KISAKHACK: pointer alignment via integer mask. Upstream used uint32_t
+    // (32-bit pointer); use uintptr_t to keep all bits on a 64-bit host.
+    g_streamPos = (uint8_t *)(~(uintptr_t)alignment & (uintptr_t)&g_streamPos[alignment]);
     return g_streamPos;
 }
